Manage Modal

+++ src/app/dashboard/dashboard.component.ts
@@ -280,7 +280,7 @@ export class DashboardComponent implements OnInit {
   removeHaribhaktList(id) {
     Swal.fire({
       title: 'Are you sure?',
-      text: "You want to delete! If you delete Offer then all the offer data will be delete.",
+      text: "You want to delete! If you delete Registerd Member then member will be removed from family.",
       icon: 'warning',
       showCancelButton: true,
       customClass: {
@@ -291,28 +291,43 @@ export class DashboardComponent implements OnInit {
       buttonsStyling: false
     }).then((result) => {
       if (result.value == true) {
-        this.dashboardService.removeHaribhaktDetails(id).subscribe((req) => {
-          this.apiService.showNotification('top', 'right', 'Offer removed Successfully.', 'success');
-
-        })
-        Swal.fire(
-          {
-            title: 'Deleted!',
-            text: 'Your Offer has been deleted.',
-            icon: 'success',
-            customClass: {
-              confirmButton: "btn btn-success",
-            },
-            buttonsStyling: false
+        Swal.fire({
+          title: 'Are you sure?',
+          text: "You want to delete! If you delete Registerd Member then member will be removed from family.",
+          icon: 'warning',
+          showCancelButton: true,
+          customClass: {
+            confirmButton: 'btn btn-success',
+            cancelButton: 'btn btn-danger',
+          },
+          confirmButtonText: 'Yes',
+          buttonsStyling: false
+        }).then((result) => {
+          if (result.value == true) {
+            this.dashboardService.removeHaribhaktDetails(id).subscribe((req) => {
+              this.apiService.showNotification('top', 'right', 'Offer removed Successfully.', 'success');
+          
+            })
+            Swal.fire(
+              {
+                title: 'Deleted!',
+                text: 'Your Offer has been deleted.',
+                icon: 'success',
+                customClass: {
+                  confirmButton: "btn btn-success",
+                },
+                buttonsStyling: false
+              }
+            )
+            this.getHaribhakt();
+          
           }
-        )
-        this.getHaribhakt();
+        })
       }
     })
 
   }
-
-
+ 
 
   getRedTickCount() {
     this.dashboardService.getRedtickCount().subscribe((res: any) => {
